move: 이동 테스트 코드
요구 사항 적용을 위한 코드 파일 병함
